### 2020-05-27 #2261 @whedon
[PRE REVIEW]: gobbli: A uniform interface to deep learning for text in Python
**Submitting author:** @jasonnance (<a href="http://orcid.org/0000-0003-4127-3198">Jason Nance</a>)
**Repository:** <a href="https://github.com/RTIInternational/gobbli" target ="_blank">https://github.com/RTIInternational/gobbli</a>
**Version:** v0.2.0
