### 2020-05-27 034ef21
Attempt to fix failing readthedocs builds
