+++ docs/requirements.txt
@@ -1,3 +1,5 @@
+# Needed for readthedocs to install everything required to build the docs
+-r ../requirements.txt
 sphinx==2.1.0
 sphinx-autobuild==0.7.1
 sphinx-autodoc-typehints==1.6.0
