### 2020-05-27 cb63505
Clean up casing in references and add DOIs where suggested by whedon bot
