+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @article{Pang:2008,
-  title={Opinion mining and sentiment analysis},
+  title={Opinion Mining and Sentiment Analysis},
   author={Pang, Bo and Lee, Lillian},
-  journal={Foundations and trends in information retrieval},
+  journal={Foundations and Trends in Information Retrieval},
   volume={2},
   number={1-2},
   pages={1--135},
@@ -10,18 +10,19 @@
 }
 
 @incollection{Aggarwal:2012,
-  title={A survey of text classification algorithms},
+  title={A Survey of Text Classification Algorithms},
   author={Aggarwal, Charu C and Zhai, ChengXiang},
-  booktitle={Mining text data},
+  booktitle={Mining Text Data},
   pages={163--222},
   year={2012},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/978-1-4614-3223-4_6}
 }
 
 @article{Weiss:2016,
   title={A survey of transfer learning},
   author={Weiss, Karl and Khoshgoftaar, Taghi M and Wang, DingDing},
-  journal={Journal of Big data},
+  journal={Journal of Big Data},
   volume={3},
   number={1},
   pages={9},
@@ -30,40 +31,42 @@
 }
 
 @incollection{Torrey:2010,
-  title={Transfer learning},
+  title={Transfer Learning},
   author={Torrey, Lisa and Shavlik, Jude},
-  booktitle={Handbook of research on machine learning applications and trends: algorithms, methods, and techniques},
+  booktitle={Handbook of Research on Machine Learning Applications and Trends: Algorithms, Methods, and Techniques},
   pages={242--264},
   year={2010},
   publisher={IGI Global}
 }
 
 @article{Rajpurkar:2016,
-  title={Squad: 100,000+ questions for machine comprehension of text},
+  title={SQuAD: 100,000+ Questions for Machine Comprehension of Text},
   author={Rajpurkar, Pranav and Zhang, Jian and Lopyrev, Konstantin and Liang, Percy},
   journal={arXiv preprint arXiv:1606.05250},
-  year={2016}
+  year={2016},
+  doi={10.18653/v1/d16-1264}
 }
 
 @article{Sun:2019,
-  title={Ernie: Enhanced representation through knowledge integration},
+  title={ERNIE: Enhanced Representation through Knowledge Integration},
   author={Sun, Yu and Wang, Shuohuan and Li, Yukun and Feng, Shikun and Chen, Xuyi and Zhang, Han and Tian, Xin and Zhu, Danxiang and Tian, Hao and Wu, Hua},
   journal={arXiv preprint arXiv:1904.09223},
   year={2019}
 }
 
 @article{Raffel:2019,
-  title={Exploring the limits of transfer learning with a unified text-to-text transformer},
+  title={Exploring the Limits of Transfer Learning with a Unified Text-to-Text Transformer},
   author={Raffel, Colin and Shazeer, Noam and Roberts, Adam and Lee, Katherine and Narang, Sharan and Matena, Michael and Zhou, Yanqi and Li, Wei and Liu, Peter J},
   journal={arXiv preprint arXiv:1910.10683},
   year={2019}
 }
 
 @article{Liu:2019,
-  title={Multi-task deep neural networks for natural language understanding},
+  title={Multi-Task Deep Neural Networks for Natural Language Understanding},
   author={Liu, Xiaodong and He, Pengcheng and Chen, Weizhu and Gao, Jianfeng},
   journal={arXiv preprint arXiv:1901.11504},
-  year={2019}
+  year={2019},
+  doi={10.18653/v1/p19-1441}
 }
 
 @article{Wolf:2019,
@@ -83,14 +86,15 @@
 }
 
 @article{Wei:2019,
[…]
