### 2020-05-27 #2261 @jasonnance
Thanks, Daniel!  We understand JOSS has a large backlog and are in no hurry.  I've tried to clean up the bib entries a bit.
