### 2020-06-12 #2261 @labarba
👋 @bmcfee -- Could you have a look at this submission and let us know if you can handle it as editor?
Title — gobbli: A uniform interface to deep learning for text in Python
[ Article proof](https://github.com/openjournals/joss-papers/blob/joss.02261/joss.02261/10.21105.joss.02261.pdf)
