### 2020-06-26 #2395 @whedon
[REVIEW]: gobbli: A uniform interface to deep learning for text in Python
**Submitting author:** @jasonnance (<a href="http://orcid.org/0000-0003-4127-3198">Jason Nance</a>)
**Repository:** <a href="https://github.com/RTIInternational/gobbli" target ="_blank">https://github.com/RTIInternational/gobbli</a>
**Version:** v0.2.0

@w4ngatang, @ljvmiranda921, & @sisco0, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
