### 2020-06-27 #2395 @W4ngatang
Hi @jasonnance,

Just to confirm, the paper is [here](https://github.com/RTIInternational/gobbli/pull/22/files)? I see that it's currently in a PR; is it ready for review, or should we wait?

Best,
Alex
