### 2020-06-30 20cad6d
Fix classification report for multiclass case
The report was outputting 0 for support for all classes due to a mismatch
between label indices and string labels
