### 2020-08-28 6e20aa3
Fix Unix-specific code to get directory creation time
