### 2021-02-17 #2395 @arfon
:wave: @nreimers and @JoPfeiff - would you be willing to step in on behalf of @thomwolf here and review this submission for JOSS?

The submission we'd be asking you to review is [_gobbli: A uniform interface to deep learning for text in Python_](https://github.com/RTIInternational/gobbli).
