### 2021-02-17 #2395 @nreimers
Hi @arfon
I am currently on parental leave and sadly will not have time in the next months for reviewing the submission of JOSS.

Best
Nils
