### 2021-03-25 3a603d1
Bump pyyaml from 5.3 to 5.4 in /benchmark
Bumps [pyyaml](https://github.com/yaml/pyyaml) from 5.3 to 5.4.
- [Release notes](https://github.com/yaml/pyyaml/releases)
- [Changelog](https://github.com/yaml/pyyaml/blob/master/CHANGES)
- [Commits](https://github.com/yaml/pyyaml/compare/5.3...5.4)

Signed-off-by: dependabot[bot] <support@github.com>
