+++ benchmark/requirements.txt
@@ -1,4 +1,4 @@
-PyYAML==5.3
+PyYAML==5.4
 tabulate==0.8.6
 umap-learn==0.3.10
 matplotlib==3.1.3
