### 2021-04-15 #2395 @sisco0
The current document shows comments on each one of the Checklist points for https://github.com/openjournals/joss-reviews/issues/2395. After reviewing the solution, I could say that it seems a promising solution for allowing researchers to integrate into current development pipelines. This tool could provide a fast way for comparing different solutions while easing the task of model configuration and data augmentation duties.
## Installation: Does installation proceed as outlined in the documentation?
The package could be installed by using pip as usual. An environment has been set up by the use of Pipenv, where a Pipfile was created at the root folder.
It would be desirable to have Poetry for the development project to take care of the dependencies. The development project has been installed as well, and CI tests could be run without any inconvenience.
## Functionality: Have the functional claims of the software been confirmed?
Source code containing augmentation algorithms as listed in the paper are enumerated below with references:
Back translation https://github.com/RTIInternational/gobbli/blob/2f4f6abb3a3a9635afb5ae625010b1ad60489bbc/gobbli/augment/marian/model.py#L14
Word replacement https://github.com/RTIInternational/gobbli/blob/2f4f6abb3a3a9635afb5ae625010b1ad60489bbc/gobbli/augment/word2vec.py
Wordnet, which represents a Contextual Augmentation implementation https://github.com/RTIInternational/gobbli/blob/2f4f6abb3a3a9635afb5ae625010b1ad60489bbc/gobbli/augment/wordnet.py#L44
The classification experiment at https://gobbli.readthedocs.io/en/latest/quickstart.html#classification-experiment was driven successfully. As a drawback, no output is returned in the console, and the application stops in less than 10 seconds. That behaviour could make people think about if the application ran successfully or not. This test confirmed the ease of a solution to be built on top of this Python package.
## Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
There are no performance claims for this project. It shows the option of running under GPU, which could be accomplished by using GPU-enabled docker configurations.

## A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
This point is clearly pointed out in the PDF, where it is stated than:
gobbli is designed to emphasize simplicity and interoperability rather than customization and performance in order to make deep learning more accessible to applied researchers.
gobbli is a Python library intended to bridge state-of-the-art research in natural language
processing and application to real-world problems.
## Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
The list of dependencies is clear for the end users, and for the developers, as different requirements.txt files hav been found out in the repository, and these are referenced in the main `README.md` file.
## Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
Different examples are shown at https://gobbli.readthedocs.io/en/latest/quickstart.html#high-level-api-experiments. They represent real problems simplifications, but the source code for real problems would be moreover the same.
## Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
API documentation is highly embedded in the current code, where the documentation is automatically generated as well. It could be seen at https://gobbli.readthedocs.io/en/latest/api.html that there exists an index showing each component from the current solution. As an example, it could be seen at https://gobbli.readthedocs.io/en/latest/auto/gobbli.augment.html#module-gobbli.augment that there is a high detail level on this part of the project. Congratulations.
## Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
CI exists and it is runnable. I could see a point against the maintenance of the project and new developers to integrate into the project that running the CI takes high memory requirements. It should be researched if CI could be lighter.
## Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support.
There exists a CONTRIBUTING.md file, which currently holds the CLA Link to be signed by contributors and show instructions on how to contribute with Code Style and Testing. In that same document, it is specified that Code Reviews would be based on Pull Requests that attend one point at a time.
There exists a template for Bug Reports at https://github.com/RTIInternational/gobbli/blob/2f4f6abb3a3a9635afb5ae625010b1ad60489bbc/.github/ISSUE_TEMPLATE/bug-report.md
I could not find any Seek Support part in the README.md, but there exist help commands on each one of the CLI solutions provided and Github Issues are reachable.

# Software paper
## Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
These points have been provided in the third paragraph as cited before.
A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
This is already pointed out at portions of the summary like:
“A practitioner may therefore be required to learn a new programming language, a deep learning library...“
“This approach allows… without spending time adapting…”
The target audience is pointed out in the phrase shown below:
“gobbli is designed to emphasize simplicity and interoperability rather than customization and performance in order to make deep learning more accessible to applied researchers.”
## State of the field: Do the authors describe how this software compares to other commonly-used packages?
They clearly state that this package could be compared to transformers and fastai.
## Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
The next suggestions are made for the correctness of the document:

- [ ] In the last few years, a concept known as transfer learning...
- [x] In the last few years, transfer learning…
- [ ] The transfer learning model thus learns a rich representation of language which can be fine-tuned
- [x] The transfer learning model thus learns a rich representation of language, which can be fine-tuned
- [ ] and a model interface whenever they want to evaluate feasibility of a new model on a custom task
- [x] and a model interface whenever they want to evaluate the feasibility of a new model on a custom task

I also propose the next change, taking into account the formality for this text:

- [ ] which is often limited in size and/or diversity
- [x] which is often limited in size and diversity
