### 2021-04-16 a708871
Apply wording fixes to paper from review
