### 2021-04-16 #2395 @jasonnance
Thanks, @sisco0!  I've implemented your suggested changes and had the pdf regenerated.
