### 2021-04-19 #2395 @ljvmiranda921
Hi everyone (@arfon @jasonnance ) ! I'm done reviewing this repository. I've checked all the boxes in the reviewer guidelines above. I **approve** this submission with my review and suggestions below:

## Review Summary

The `gobbli` library provides a high-level NLP interface that eases the use and speeds-up experimentation in the field. Aside from abstracting away other libraries such as spacy, fastai, and transformers, it also provides helpful utilities for data augmentation, model evaluation, and exploration.

The paper is well-written, with a clear statement of need. It was also able to differentiate against other NLP libraries like `huggingface/transformers`, `explosion/spacy`, etc. that are commonly-found today. I see this as an all-in-one tool for NLP needs, and maybe useful in both research and production. In addition, I appreciate the inclusion of a streamlit UI for most-common NLP tasks, it simplifies the setup needed for researchers and practitioners alike.

The repository is well organized, with clear documentation and tests. It exists under the OSI-approved Apache-2.0 License with Contributing guidelines for the community. I tested all the functional claims from the README in my machine (Ubuntu 20.04) and they worked as expected.

## Suggestions

This may not be necessary for publication (of course, with the editor's discretion). But may be useful for the improvement and sustainability of the project in the future:

- [ ] Ensure that the CI checks pass. As of review time (April 19, 18:09 GMT+8), it's erroring-out on Travis.
- [ ] Instead of having multiple bash scripts, I suggest using a `Makefile` for generating docs, running tests, CI, etc. It gives you an all-in-one recipe for doing multiple tasks
- [ ] Explicitly state the linting tool used in [your Contributing doc](https://github.com/RTIInternational/gobbli/blob/master/CONTRIBUTING.md) so that's easier for contributors to run the formatter even before making a PR. I saw that you're using `black`. You can also opt to use pre-commits to automate things further.

In my opinion, the ones above aren't prerequisites for acceptance.
