### 2021-06-22 80eb403
Fix wrong type returned by embeddings for USE model
