### 2021-06-22 00fdf67
Correct type annotations for USE embeddings
