### 2021-06-23 ff2d1d3
Apply fixes for CI build errors
The old versions of the transformers library were pulling in the latest
version of sentencepiece, which apparently added some build
requirements.  Fixed to an older version which should still be
compatible.
