+++ gobbli/model/spacy/src/requirements.txt
@@ -6,6 +6,7 @@ pandas==0.25.0
 # https://github.com/explosion/spacy-transformers/pull/120
 spacy==2.2.1
 spacy-transformers==0.5.1
+sentencepiece==0.1.86
 
 # We're using the PyTorch image with CUDA 10.1, but spaCy doesn't have an extra
 # requirements specifier for CUDA 10.1 at the time of this writing (it only has 10.0).
