### 2021-06-23 #2395 @jasonnance
Great, thanks @arfon.  We've released v0.2.4.  We had existing automation set up with Zenodo to create DOIs for each new version, but the names don't match the paper title, although the authors are the same.  Do we need to create a new version of the DOI for v0.2.4 and edit the title to match the paper title?  Or should we create an entirely separate DOI for this paper submission?  Here's what we currently have: https://doi.org/10.5281/zenodo.3406400
